requirements.txt
base58==2.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colander==1.7.0
ecdsa==0.15
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
libeospy==2.0.1
MarkupSafe==1.1.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
translationstring==1.3
urllib3==1.25.9
Werkzeug==1.0.1
